moved mgcncs generators to generators dir
